renaming; update example schemas
