+++ code/aind_auto_train/auto_train_manager.py
@@ -8,7 +8,7 @@ import pandas as pd
 
 from aind_auto_train.schema.curriculum import TrainingStage
 from aind_auto_train.schema.task import metrics_class, Metrics, DynamicForagingMetrics
-from aind_auto_train.curriculums.coupled_baiting import curriculum as coupled_baiting_curriculum
+from aind_auto_train.curriculums.coupled_baiting_0p1 import curriculum as coupled_baiting_curriculum
 from aind_auto_train.util.aws_util import import_df_from_s3, export_df_to_s3
 from aind_auto_train.plot.manager import plot_manager_all_progress
 from aind_auto_train.curriculum_manager import CurriculumManager
