+++ code/aind_auto_train/curriculums/Coupled Baiting_curriculum_v0.1_schema_v0.3.json
@@ -14,6 +14,7 @@
             "RewardFamily": 3,
             "RewardPairsN": 1,
             "UncoupledReward": "",
+            "Randomness": "Exponential",
             "BlockMin": 10,
             "BlockMax": 20,
             "BlockBeta": 5,
@@ -21,6 +22,7 @@
             "DelayMin": 0.5,
             "DelayMax": 0.5,
             "DelayBeta": 0.0,
+            "RewardDelay": 0.0,
             "AutoReward": true,
             "AutoWaterType": "Natural",
             "Multiplier": 0.5,
@@ -51,9 +53,7 @@
             "GiveWaterR_volume": 3.0,
             "IncludeAutoReward": false,
             "SaveTraining": true,
-            "InitiallyInactiveN": 2,
-            "Randomness": "Exponential",
-            "qt_spinbox_lineedit": 5.0
+            "InitiallyInactiveN": 2
         },
         "Stage 2": {
             "training_stage": "Stage 2",
@@ -64,6 +64,7 @@
             "RewardFamily": 1,
             "RewardPairsN": 1,
             "UncoupledReward": "",
+            "Randomness": "Exponential",
             "BlockMin": 10,
             "BlockMax": 40,
             "BlockBeta": 10,
@@ -71,6 +72,7 @@
             "DelayMin": 1.0,
             "DelayMax": 1.0,
             "DelayBeta": 0.0,
+            "RewardDelay": 0.0,
             "AutoReward": true,
             "AutoWaterType": "Natural",
             "Multiplier": 0.5,
@@ -101,9 +103,7 @@
             "GiveWaterR_volume": 3.0,
             "IncludeAutoReward": false,
             "SaveTraining": true,
-            "InitiallyInactiveN": 2,
-            "Randomness": "Exponential",
-            "qt_spinbox_lineedit": 5.0
+            "InitiallyInactiveN": 2
         },
         "Stage 3": {
             "training_stage": "Stage 3",
@@ -114,6 +114,7 @@
             "RewardFamily": 1,
             "RewardPairsN": 1,
             "UncoupledReward": "",
+            "Randomness": "Exponential",
             "BlockMin": 20,
             "BlockMax": 60,
             "BlockBeta": 20,
@@ -121,6 +122,7 @@
             "DelayMin": 1.5,
             "DelayMax": 1.5,
             "DelayBeta": 0.0,
+            "RewardDelay": 0.0,
             "AutoReward": true,
             "AutoWaterType": "Natural",
             "Multiplier": 0.5,
@@ -151,9 +153,7 @@
             "GiveWaterR_volume": 3.0,
             "IncludeAutoReward": false,
             "SaveTraining": true,
-            "InitiallyInactiveN": 2,
-            "Randomness": "Exponential",
-            "qt_spinbox_lineedit": 5.0
+            "InitiallyInactiveN": 2
         },
         "Stage final": {
             "training_stage": "Stage final",
@@ -164,6 +164,7 @@
             "RewardFamily": 1,
             "RewardPairsN": 4,
             "UncoupledReward": "",
+            "Randomness": "Exponential",
             "BlockMin": 20,
             "BlockMax": 60,
             "BlockBeta": 20,
@@ -171,6 +172,7 @@
             "DelayMin": 2.0,
             "DelayMax": 2.0,
             "DelayBeta": 0,
+            "RewardDelay": 0.0,
             "AutoReward": false,
             "AutoWaterType": "Natural",
             "Multiplier": 0.5,
@@ -201,9 +203,7 @@
             "GiveWaterR_volume": 3.0,
             "IncludeAutoReward": false,
             "SaveTraining": true,
-            "InitiallyInactiveN": 2,
-            "Randomness": "Exponential",
-            "qt_spinbox_lineedit": 5.0
+            "InitiallyInactiveN": 2
         }
     },
     "curriculum": {
