+++ code/aind_auto_train/curriculums/coupled_baiting_0p1.py
