+++ code/aind_auto_train/curriculums/coupled_baiting_0p2.py
@@ -13,10 +13,10 @@ from aind_auto_train.curriculum_manager import LOCAL_SAVED_CURRICULUM_ROOT
 
 
 # Reuse the curriculum from 0.1
-from aind_auto_train.curriculums.coupled_baiting import curriculum as curriculum_0_1
+from aind_auto_train.curriculums.coupled_baiting_0p1 import curriculum as curriculum_0p1
 
 # Override some parameters
-curriculum = curriculum_0_1.model_copy()
+curriculum = curriculum_0p1.model_copy()
 curriculum.curriculum_version = "0.2"
 curriculum.curriculum_description = "More stringent criteria before GRADUATED than 0.1"
